[Fix #869] Splitting ExpressionFactory from ModelFactory
