--- experimental/agentic/src/main/java/io/serverlessworkflow/impl/expressions/agentic/AgenticExpressionFactory.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2020-Present The Serverless Workflow Specification Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.serverlessworkflow.impl.expressions.agentic;
-
-import io.serverlessworkflow.impl.WorkflowModelFactory;
-import io.serverlessworkflow.impl.expressions.func.JavaExpressionFactory;
-
-public class AgenticExpressionFactory extends JavaExpressionFactory {
-
-  private final WorkflowModelFactory modelFactory = new AgenticModelFactory();
-
-  @Override
-  public WorkflowModelFactory modelFactory() {
-    return modelFactory;
-  }
-
-  public int priority() {
-    return DEFAULT_PRIORITY - 1;
-  }
-}
