--- experimental/agentic/src/main/resources/META-INF/services/io.serverlessworkflow.impl.expressions.ExpressionFactory
@@ -1,1 +1,0 @@
-io.serverlessworkflow.impl.expressions.agentic.AgenticExpressionFactory
