+++ experimental/lambda/pom.xml
@@ -14,7 +14,16 @@
         </dependency>
         <dependency>
             <groupId>io.serverlessworkflow</groupId>
+            <artifactId>serverlessworkflow-experimental-model</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>io.serverlessworkflow</groupId>
             <artifactId>serverlessworkflow-impl-core</artifactId>
+        </dependency>
+         <dependency>
+            <groupId>io.serverlessworkflow</groupId>
+            <artifactId>serverlessworkflow-impl-jq</artifactId>
         </dependency>
         <dependency>
             <groupId>io.serverlessworkflow</groupId>
