--- impl/jackson/src/main/java/io/serverlessworkflow/impl/expressions/jq/JacksonModelDeserializer.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2020-Present The Serverless Workflow Specification Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.serverlessworkflow.impl.expressions.jq;
-
-import com.fasterxml.jackson.core.JacksonException;
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.deser.std.StdDeserializer;
-import java.io.IOException;
-
-public class JacksonModelDeserializer extends StdDeserializer<JacksonModel> {
-
-  private static final long serialVersionUID = 1L;
-
-  protected JacksonModelDeserializer() {
-    super(JacksonModel.class);
-  }
-
-  @Override
-  public JacksonModel deserialize(JsonParser p, DeserializationContext ctxt)
-      throws IOException, JacksonException {
-    return new JacksonModel(p.readValueAsTree());
-  }
-}
