--- impl/jackson/src/main/java/io/serverlessworkflow/impl/expressions/jq/JacksonModelSerializer.java
@@ -1,36 +1,0 @@
-/*
- * Copyright 2020-Present The Serverless Workflow Specification Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.serverlessworkflow.impl.expressions.jq;
-
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.databind.SerializerProvider;
-import com.fasterxml.jackson.databind.ser.std.StdSerializer;
-import java.io.IOException;
-
-public class JacksonModelSerializer extends StdSerializer<JacksonModel> {
-
-  private static final long serialVersionUID = 1L;
-
-  protected JacksonModelSerializer() {
-    super(JacksonModel.class);
-  }
-
-  @Override
-  public void serialize(JacksonModel value, JsonGenerator gen, SerializerProvider provider)
-      throws IOException {
-    gen.writeTree(value.node);
-  }
-}
