+++ impl/jackson/src/main/resources/META-INF/services/io.serverlessworkflow.impl.WorkflowModelFactory
@@ -1,0 +1,1 @@
+io.serverlessworkflow.impl.model.jackson.JacksonModelFactory
